refactor(bedrock): クリーンアーキテクチャに従ったBedrock構造のリファクタリング
- Infrastructure層: BedrockClientでAPI通信のみを担当
- Domain層: ResearchDomainServiceにコンテンツ処理を統合
- Application層: ExecuteResearchUseCaseでフロー制御を簡素化
- ProcessContentUseCaseは不要（他のUseCaseから呼ばれる処理のため）
- 350行のBedrockContentProcessorを責務別に分離
- PerplexityとBedrock両方でクリーンアーキテクチャの一貫性を確保

Refs: #3
