+++ src/app/api/research/route.ts
@@ -1,11 +1,7 @@
 import { NextRequest, NextResponse } from "next/server";
 import { createExecuteResearchUseCase } from "@/shared/useCases";
-import { BedrockContentProcessor } from "@/shared/infrastructure/external/bedrock/ContentProcessor";
 import { CreateResearchRequest } from "@/shared/api/generated/models";
-import type {
-  VoicePattern,
-  ResearchResult,
-} from "@/shared/api/generated/models";
+import type { VoicePattern } from "@/shared/api/generated/models";
 import { executeResearchBody } from "@/shared/api/generated/zod";
 
 /**
@@ -50,58 +46,15 @@ export async function POST(request: NextRequest) {
       );
     }
 
-    // 1. リサーチユースケースの実行
+    // リサーチユースケースの実行（コンテンツ処理も含む）
     const executeResearchUseCase = createExecuteResearchUseCase(apiKey);
-    const initialResult = await executeResearchUseCase.execute({
+    const result = await executeResearchUseCase.execute({
       query: validation.data.query,
       selectedText: validation.data.selectedText,
       voiceCommand: validation.data.voiceCommand as VoicePattern, // 型アサーション
     });
 
-    // 2. BedrockでHTML変換を実行（各results.contentに対して）
-    try {
-      const bedrockProcessor = new BedrockContentProcessor();
-
-      const enhancedResults = await Promise.all(
-        (initialResult.results || []).map(async (result: ResearchResult) => {
-          try {
-            // BedrockでHTML変換を実行
-            const enhancedContent = await bedrockProcessor.processContent(
-              result.content,
-              initialResult.citations || [],
-              initialResult.searchResults || [],
-            );
-
-            return {
-              ...result,
-              content: enhancedContent.htmlContent,
-              processedCitations: enhancedContent.processedCitations,
-            };
-          } catch (error) {
-            console.warn(
-              `Bedrock processing failed for result ${result.id}:`,
-              error,
-            );
-            // フォールバック: 元のコンテンツを使用
-            return result;
-          }
-        }),
-      );
-
-      const finalResult = {
-        ...initialResult,
-        results: enhancedResults,
-      };
-
-      return NextResponse.json(finalResult, { status: 200 });
-    } catch (error) {
-      console.warn(
-        "Bedrock enhancement failed, returning basic result:",
-        error,
-      );
-      // フォールバック: 基本的なHTML変換結果を返す
-      return NextResponse.json(initialResult, { status: 200 });
-    }
+    return NextResponse.json(result, { status: 200 });
   } catch (error) {
     console.error("Research API error:", error);
